+++ mobile/src/main/java/com/cscao/apps/mobirnn/model/Model.java
@@ -12,12 +12,12 @@ import java.io.IOException;
 
 public class Model {
 
-    private double[][][] weights;
-    private double[][] biases;
-    private double[][] w_in;
-    private double[][] w_out;
-    private double[] b_in;
-    private double[] b_out;
+    private float[][][] weights;
+    private float[][] biases;
+    private float[][] w_in;
+    private float[][] w_out;
+    private float[] b_in;
+    private float[] b_out;
     private int layerSize;
     private int hidden_units;
     private boolean isModeCpu = true;
@@ -52,19 +52,19 @@ public class Model {
         if (layerSize != biasesPath.length) throw new AssertionError();
         this.layerSize = layerSize;
 
-        double[] b_in = DataUtil.parseBias(bInPath);
-        double[] b_out = DataUtil.parseBias(bOutPath);
+        float[] b_in = DataUtil.parseBias(bInPath);
+        float[] b_out = DataUtil.parseBias(bOutPath);
         this.hidden_units = b_in.length;
 
-        double[][] w_in = DataUtil.parseWeight(wInPath);
-        double[][] w_out = DataUtil.parseWeight(wOutPath);
+        float[][] w_in = DataUtil.parseWeight(wInPath);
+        float[][] w_out = DataUtil.parseWeight(wOutPath);
 
-        double[][][] weights = new double[layerSize][][];
+        float[][][] weights = new float[layerSize][][];
         for (int i = 0; i < layerSize; i++) {
             weights[i] = DataUtil.parseWeight(dataPath + weightsPath[i]);
         }
 
-        double[][] biases = new double[layerSize][];
+        float[][] biases = new float[layerSize][];
         for (int i = 0; i < layerSize; i++) {
             biases[i] = DataUtil.parseBias(dataPath + biasesPath[i]);
         }
@@ -77,15 +77,15 @@ public class Model {
         this.b_out = b_out;
     }
 
-    private double[][] calcCellOneStep(double[] in_, double[] c_, double[] h_, int layer) {
+    private float[][] calcCellOneStep(float[] in_, float[] c_, float[] h_, int layer) {
         // concat in_ and h_ and do xw_plu_b
-        double[] concat = Matrix.concat(in_, h_);
-        double[] linearResult = Matrix.vecAddVec(Matrix.vecMulMat(concat, this.weights[layer]),
+        float[] concat = Matrix.concat(in_, h_);
+        float[] linearResult = Matrix.vecAddVec(Matrix.vecMulMat(concat, this.weights[layer]),
                 this.biases[layer]);
-        double[] i = Matrix.split(linearResult, 4, 0);
-        double[] j = Matrix.split(linearResult, 4, 1);
-        double[] f = Matrix.split(linearResult, 4, 2);
-        double[] o = Matrix.split(linearResult, 4, 3);
+        float[] i = Matrix.split(linearResult, 4, 0);
+        float[] j = Matrix.split(linearResult, 4, 1);
+        float[] f = Matrix.split(linearResult, 4, 2);
+        float[] o = Matrix.split(linearResult, 4, 3);
 
         int size = c_.length;
         for (int k = 0; k < size; k++) {
@@ -93,13 +93,13 @@ public class Model {
             h_[k] = DataUtil.tanh(c_[k]) * DataUtil.sigmod(o[k]);
         }
 
-        double[][] state = new double[2][];
+        float[][] state = new float[2][];
         state[0] = c_;
         state[1] = h_;
         return state;
     }
 
-    public int predict(double[][] x) {
+    public int predict(float[][] x) {
         if (isModeCpu) {
             return predictOnCpu(x);
         } else {
@@ -107,35 +107,35 @@ public class Model {
         }
     }
 
-    private int predictOnCpu(double[][] x) {
+    private int predictOnCpu(float[][] x) {
         int timeSteps = x.length;
 
-        double[][] outputs = new double[timeSteps][];
+        float[][] outputs = new float[timeSteps][];
         x = DataUtil.relu(Matrix.addVec(Matrix.multiply(x, w_in), b_in));
 
         for (int j = 0; j < layerSize; j++) {
-            double[] c = new double[hidden_units];
-            double[] h = new double[hidden_units];
+            float[] c = new float[hidden_units];
+            float[] h = new float[hidden_units];
 
             for (int k = 0; k < timeSteps; k++) {
-                double[][] state = calcCellOneStep(x[k], c, h, j);
+                float[][] state = calcCellOneStep(x[k], c, h, j);
                 c = state[0];
                 h = state[1];
                 System.arraycopy(h, 0, x[k], 0, hidden_units);
-                outputs[k] = new double[hidden_units];
+                outputs[k] = new float[hidden_units];
                 System.arraycopy(h, 0, outputs[k], 0, hidden_units);
 //                x[k] = h;
 //                outputs[k] = h;
             }
         }
 
-        double[] outProb = Matrix.vecAddVec(Matrix.vecMulMat(outputs[timeSteps - 1], w_out), b_out);
+        float[] outProb = Matrix.vecAddVec(Matrix.vecMulMat(outputs[timeSteps - 1], w_out), b_out);
 //        System.out.println("out:" + Arrays.toString(outProb).replaceAll("[\\[ | \\] | ,]", " "));
         return DataUtil.argmax(outProb) + 1;
     }
 
     // TODO: 4/7/17Friday using renderscript to implement gpu code
-    private int predictOnGpu(double[][] x) {
+    private int predictOnGpu(float[][] x) {
         return 0;
     }
 }
